Initial Comit
Prior to the integration of the front end, all components work as
desired (using a basic coffee.html testing platform).
The site allows a user to purchase a coffee, pay with PayPal, get the
information back from PayPal, add this data to a private database, then
allow a page to pull from said database. Once an order is done, a
barista can hit done which will send a text to the customer and remove
the item from the database.